pyproject.toml
[tool.poetry]
name = "pycrank"
packages = [
  { include = "crank.py" },
]
version = "0.1.0"
description = "Simple static-site generator"
authors = ["Jonathan Bowman <git@bowmanjd.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
"paka.cmark" = "^2.1"
selectolax = "^0.1.13"
pygments = "^2.4"

[tool.poetry.dev-dependencies]
ptpython = "^2.0"
pytest = "^4.6"

[tool.poetry.scripts]
crank = 'crank:run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
